pyproject.toml
[project]
name = "kaspi_parser"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Sabyr Shatekov", email = "shatekov.sabyr@gmail.com"},
]
dependencies = ["fastapi>=0.115.5", "uvicorn>=0.32.0", "pydantic>=2.9.2", "python-dotenv>=1.0.1", "pytest>=8.3.3"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
